Change profile image

+++ internal/persistence/s3.go
@@ -16,7 +16,7 @@ import (
 )
 
 var Bucket = os.Getenv("DO_SPACES_BUCKET")
-var BaseUrl = os.Getenv("DO_SPACES_BASE_URL")
+var S3BaseUrl = os.Getenv("DO_SPACES_BASE_URL")
 var S3 *s3.Client
 var DealerImagesFolder = "dealer-images"
 var DealImagesFolder = "deal-images"
@@ -46,7 +46,7 @@ func InitS3() {
 
 	S3 = s3.NewFromConfig(cfg)
 
-	zap.L().Sugar().Infof("S3 init done: region=%s, endpoint=%s, bucket=%s, baseUrl=%s", region, endpoint, Bucket, BaseUrl)
+	zap.L().Sugar().Infof("S3 init done: region=%s, endpoint=%s, bucket=%s, baseUrl=%s", region, endpoint, Bucket, S3BaseUrl)
 }
 
 func UploadImage(path string, image *multipart.FileHeader) error {
@@ -61,6 +61,7 @@ func UploadImage(path string, image *multipart.FileHeader) error {
 	if err != nil {
 		return err
 	}
+	defer file.Close()
 
 	_, err = S3.PutObject(context.TODO(), &s3.PutObjectInput{
 		Bucket:      &Bucket,
@@ -85,7 +86,7 @@ func GetImageUrls(path string) ([]string, error) {
 
 	var imageUrls []string
 	for _, content := range output.Contents {
-		imageUrls = append(imageUrls, fmt.Sprintf("%s/%s", BaseUrl, *content.Key))
+		imageUrls = append(imageUrls, fmt.Sprintf("%s/%s", S3BaseUrl, *content.Key))
 	}
 
 	return imageUrls, nil
@@ -109,6 +110,10 @@ func GetImageUrl(path string) (string, error) {
 }
 
 func DeleteImage(path string) error {
+	if len(path) == 0 {
+		return nil
+	}
+
 	zap.L().Sugar().Debugf("delete image: %s", path)
 	_, err := S3.DeleteObject(context.TODO(), &s3.DeleteObjectInput{
 		Bucket: &Bucket,
